+++ deltafi-common/src/main/java/org/deltafi/common/storage/s3/minio/MinioObjectStorageService.java
@@ -19,7 +19,8 @@ package org.deltafi.common.storage.s3.minio;
 
 import io.minio.*;
 import io.minio.errors.*;
-import io.minio.messages.*;
+import io.minio.messages.DeleteError;
+import io.minio.messages.DeleteObject;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.input.CountingInputStream;
@@ -31,7 +32,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
-import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -39,80 +39,9 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @Slf4j
 public class MinioObjectStorageService implements ObjectStorageService {
-    private static final String AGE_OFF = "AgeOff";
 
     protected final MinioClient minioClient;
     protected final MinioProperties minioProperties;
-
-    @Override
-    public boolean bucketExists(String bucketName) throws ObjectStorageException {
-        try {
-            return minioClient.bucketExists(BucketExistsArgs.builder().bucket(bucketName).build());
-        } catch (Exception e) {
-            log.error("Unable to check bucket existence");
-            throw new ObjectStorageException("Unable to check bucket existence", e);
-        }
-    }
-
-    @Override
-    public void createBucket(String bucketName) throws ObjectStorageException {
-        if (!bucketExists(bucketName)) {
-            try {
-                minioClient.makeBucket(MakeBucketArgs.builder().bucket(bucketName).build());
-                log.info("Created the bucket: " + bucketName);
-            } catch (Exception e) {
-                log.error("Unable to create bucket");
-                throw new ObjectStorageException("Unable to create bucket", e);
-            }
-        }
-    }
-
-    @Override
-    public boolean expectedConfiguration(String bucketName) throws ObjectStorageException {
-        try {
-            return checkConfiguration(
-                    minioClient.getBucketLifecycle(GetBucketLifecycleArgs.builder()
-                            .bucket(bucketName).build()));
-        } catch (Exception e) {
-            log.error("Unable to get bucket lifecycle");
-            throw new ObjectStorageException("Unable to get bucket lifecycle", e);
-        }
-    }
-
-    private boolean checkConfiguration(LifecycleConfiguration configuration) {
-        return configuration != null &&
-                configuration.rules() != null &&
-                configuration.rules().size() == 1 &&
-                configuration.rules().get(0).status() == Status.ENABLED &&
-                configuration.rules().get(0).id().equals(AGE_OFF) &&
-                configuration.rules().get(0).expiration().days() == minioProperties.getExpirationDays();
-    }
-
-    @Override
-    public void setExpiration(String bucketName) throws ObjectStorageException {
-        try {
-            minioClient.setBucketLifecycle(SetBucketLifecycleArgs.builder()
-                    .bucket(bucketName)
-                    .config(getLifeCycleConfig()).build());
-            log.info("Set bucket age-off days: " + minioProperties.getExpirationDays());
-        } catch (Exception e) {
-            log.error("Unable to set bucket lifecycle");
-            throw new ObjectStorageException("Unable to set bucket lifecycle", e);
-        }
-    }
-
-    private LifecycleConfiguration getLifeCycleConfig() {
-        return new LifecycleConfiguration(List.of(
-                new LifecycleRule(
-                        Status.ENABLED,
-                        null,
-                        new Expiration((ZonedDateTime) null, minioProperties.getExpirationDays(), null),
-                        new RuleFilter(""),
-                        AGE_OFF,
-                        null,
-                        null,
-                        null)));
-    }
 
     @Override
     public InputStream getObject(ObjectReference objectReference) throws ObjectStorageException {