+++ deltafi-common/src/main/java/org/deltafi/common/storage/s3/minio/MinioProperties.java
@@ -27,5 +27,4 @@ public class MinioProperties {
     String accessKey;
     String secretKey;
     long partSize;
-    int expirationDays;
 }